added message encryption
I included CAST5 message encryption. This includes a key generated
(currently a 128-bit key of all zeros) in NewSession() and
Message.Content encryption during SendMessage().

An important thing to note about the encryption done during
SendMessage() is that the string is first wrapped in bencode, then
padded with zero-bytes. (The padding is necessary because CAST5 is a
block cipher.)