added rsa key generation
I added support for (and requirement for) RSA keypairs. The RSAKey
interface allows clients to define their own means of reading the key
into memory. This means that the key could be stored on the disk, or
encrypted, so long as a valid *rsa.PrivateKey.

I also updated the test client to use this.

ref #2