changed SID to type uint64
I changed Session.SID and Message.SID both to type uint64, which is an
equivalent amount of data as they previously held (len 8 []byte,) but
usable as keys in map types. This means that it's very much possible to
keep a map[uint64]*Session and look up Sessions by Message.SIDs.